pyproject.toml
[tool.poetry]
name = "vault-actions"
version = "0.1.0"
description = ""
authors = ["jr200"]
license = "LICENSE"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
hvac = "^2.3.0"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
